fix(logs): move folder expansion to server-only module to fix client bundle build
